Fix list nodes when offline_cause_reason not a dict
The offline_cause is typically a dictionary, but
occasionally it's a basic string such as '17.1'.
Although json_loads can successfully parse it,
attempting to obtain keys in the rest of the code
results in errors.

This patch checks therefore that we are dealing with a
dict before trying to retrieve any key.

Signed-off-by: Fernando Royo <froyo@redhat.com>